Extract correlation id generation
